refactor: refactor naiming and separate function
