+++ .eslintrc.cjs
@@ -41,5 +41,7 @@ module.exports = {
     'import/prefer-default-export': 'off',
     'import/no-unresolved': 'off',
     'no-constructor-return': 'off',
+    'no-restricted-syntax': 'off',
+    'no-param-reassign': 'off',
   },
 };
