+++ src/controller/controllerType.ts
@@ -1,10 +1,10 @@
 import { ClientReqData, Socket } from '../types/types/common.js';
-import { regType } from '../handlers/reqType.js';
+import { registration } from '../handlers/registration.js';
 import { RegObject } from '../types/interface/reg.js';
 import { RoomObject } from '../types/interface/room.js';
-import { roomType } from '../handlers/roomType.js';
+import { createRoom } from '../handlers/createRoom.js';
 import { AddUserToRoomClient } from '../types/interface/addUser.js';
-import { addGame } from '../handlers/startGame.js';
+import { startGame } from '../handlers/startGame.js';
 import { AddShips } from '../types/interface/addShips.js';
 import { addShips } from '../handlers/addShips.js';
 import { AttackClient } from '../types/interface/attack.js';
@@ -13,9 +13,9 @@ import { RandomAttackClient } from '../types/interface/randomAttack.js';
 import { randomAttack } from '../handlers/randomAttack.js';
 
 export const controllerType = {
-  reg: (object: ClientReqData, socket: Socket) => regType(object as RegObject, socket),
-  create_room: (object: ClientReqData, socket: Socket) => roomType(object as RoomObject, socket),
-  add_user_to_room: (object: ClientReqData, socket: Socket) => addGame(object as AddUserToRoomClient, socket),
+  reg: (object: ClientReqData, socket: Socket) => registration(object as RegObject, socket),
+  create_room: (object: ClientReqData, socket: Socket) => createRoom(object as RoomObject, socket),
+  add_user_to_room: (object: ClientReqData, socket: Socket) => startGame(object as AddUserToRoomClient, socket),
   add_ships: (object: ClientReqData, socket: Socket) => addShips(object as AddShips, socket),
   attack: (object: ClientReqData, socket: Socket) => attackShips(object as AttackClient, socket),
   randomAttack: (object: ClientReqData, socket: Socket) => randomAttack(object as RandomAttackClient, socket),
