+++ src/handlers/addShips.ts
@@ -2,23 +2,15 @@ import { PositionShipsObject } from '../types/interface/position.js';
 import { AddShips } from '../types/interface/addShips.js';
 import { Socket } from '../types/types/common.js';
 import { TypeData } from '../types/enum/typeData.js';
-import { gameRoomsBase } from '../store/gameRoomsController.js';
-import { socketBase } from '../store/socketController.js';
+import { socketBase, gameRoomsBase } from '../store/index.js';
+import { updateDataShips } from '../helpers/updateDataShips.js';
+import { getResponseObject } from '../helpers/createrObjects.js';
 
 const MAX_AMOUNT_PLAYERS_IN_ROOM = 2;
 
-export const updateShips = (dataShips: PositionShipsObject) => {
-  const { ships } = dataShips;
-  const newShips = ships.map((item) => {
-    const hitpoint = item.length;
-    return { ...item, hitpoint };
-  });
-  return { ...dataShips, ships: newShips };
-};
-
 export const addShips = (object: AddShips, _socket: Socket) => {
   const objectData = JSON.parse(object.data) as PositionShipsObject;
-  const upgradeShips = updateShips(objectData);
+  const upgradeShips = updateDataShips(objectData);
   gameRoomsBase.setDataGame(objectData.gameId, upgradeShips);
   const amountPlayersReady = gameRoomsBase.getSizePlayers(objectData.gameId);
   if (amountPlayersReady === MAX_AMOUNT_PLAYERS_IN_ROOM) {
@@ -28,19 +20,9 @@ export const addShips = (object: AddShips, _socket: Socket) => {
         const { indexPlayer, ships } = item;
         const socketUser = socketBase.getSocketUser(indexPlayer);
         socketUser?.send(
-          JSON.stringify({
-            type: TypeData.START_GAME,
-            data: JSON.stringify({ ships, currentPlayerIndex: indexPlayer }),
-            id: 0,
-          })
+          getResponseObject(TypeData.START_GAME, JSON.stringify({ ships, currentPlayerIndex: indexPlayer }))
         );
-        socketUser?.send(
-          JSON.stringify({
-            type: TypeData.TURN,
-            data: JSON.stringify({ currentPlayer: objectData.indexPlayer }),
-            id: 0,
-          })
-        );
+        socketUser?.send(getResponseObject(TypeData.TURN, JSON.stringify({ currentPlayer: objectData.indexPlayer })));
       });
     }
   }
