+++ src/handlers/attackShips.ts
@@ -1,7 +1,6 @@
-// import { dataBase } from '../data_base/data_base.js';
 import { AttackClient, AttackData } from '../types/interface/attack.js';
 import { Socket } from '../types/types/common.js';
-import { checkMovePlayer, checkShootPlace, getDataPlaceShoot } from './supportAttack.js';
+import { checkMovePlayer, checkShootPlace, getDataPlaceShoot } from '../helpers/supportAttack.js';
 import { shoot } from '../helpers/shoot.js';
 import { gameRoomsBase } from '../store/index.js';
