+++ src/helpers/supportAttack.ts
